refactor: moved main state from Main to App
